Create a new rails app called account-api
Following what [our docs][0] suggest, the app was created with the
following flags to skip uncommon dependencies that we don't need:
- `--skip-javascript`
- `--skip-test`
- `--skip-bundle`
- `--skip-spring`
- `--skip-action-cable`
- `--skip-action-mailer`
- `--skip-active-storage`

We know this will only be an api so I also used the flags `--api` and
`--skip-turbolinks`. We also know that we want to use postgresql as our
database so I used the flag `-d postgresql`.

The full command used was the following:
```
rails new account-api --api --skip-javascript --skip-test \
--skip-bundle --skip-spring --skip-action-cable --skip-action-mailer \
--skip-active-storage --skip-turbolinks -d postgresql
```

[0]: https://docs.publishing.service.gov.uk/manual/setting-up-new-rails-app.html

Trello card: https://trello.com/c/ZeAKl2SM/651-set-up-a-new-app