Add endpoints to manage the Transition Checker email subscription
By putting these endpoints in this app, we can fully remove the OIDC /
OAuth logic from the Transition Checker.

The name of the controller is really specific because we've not yet
got a generic way to manage email subscriptions for multiple apps, and
will need to think about how to do that if it's something we want.