Add endpoints to manage attributes
Currently these just delegate to the govuk-attribute-service-prototype
running on the PaaS, but at some point soon we will begin storing
attributes in this app's database too.  We might migrate the
Transition Checker attributes, or we might just retire the checker -
that depends on how long it'll be staying around, which is a product
decision.

For the update endpoint, the attribute values are given as
JSON-encoded strings.  I think this is reasonable, because there isn't
really a standard for how complex nested data gets turned into HTTP
field data.  We can do the `to_json`ing in gds-api-adapters, so that
other apps don't need to worry about it.