Write Pact provider configuration
Pact provider configuration specifies which contracts are
honoured (currently only gds-api-adapters) and defines a set of
"provider states", which are a sort of precondition that the consumer
tests expect.

A Pact test has two parts:

1. The consumer test, which specifies which provider state is
expected, gives the request, and an expected response.

2. The provider configuration, which defines all of the provider
states.

There are no actual tests in the provider: rather, the requests
defined by the consumer tests are run against the provider, and pact
checks that the responses match.