Send application/json requests in the request specs
By default, Rails request specs send parameters as
application/x-www-form-urlencoded, but this has two problems:

1. gds-api-adapters sends JSON requests, so sending form data means
our test & production behaviours are very different.

2. form-data is bad at encoding complex nested data (or even just
numbers - which are decoded as strings), so we limit our app to only
what form-data supports.

A concrete problem caused by (2) is that the PATCH attributes endpoint
requires attribute values to be passed in as strings of JSON-encoded
data.  This is weird and confusing, and not at all like publishing-api
or email-alert-api.  It would be good to fix that, but first we need
these tests to do the right thing.

By instead sending JSON data in our tests (which publishing-api does
in its tests), we make our tests more closely match production
behaviour.