Allow both encoded and non-encoded attribute values
I've verified that this is backwards compatible by running the pact
tests.

The create-state endpoint accepts plain JSON, rather than encoded
values.  This commit makes the update-attributes endpoint also accept
plain JSON, to be consistent.

To support backwards compatibility, encoded attribute values are also
accepted.  If all of the attribute values are JSON strings, then the
assumption is made that they have been encoded by the sender.  In
practice, this is the case, because gds-api-adapters is the only thing
which uses this endpoint so far and that's what it does.

After this has been deployed we can:

1. update gds-api-adapters to *not* encode the attribute values
2. deploy a finder-frontend with the updated gds-api-adapters
3. remove the backwards compatibility code here

Moving to accept plain JSON values (rather than strings of encoded
values) makes this endpoint consistent with other GOV.UK API apps.