Amend Rails session headers
[Trello ticket][1]

In [frontend#2668][2] we worked out that Rails makes request headers available as

```
request.headers["HTTP_" + header_name.gsub(/-/,"_").upcase]
```

So in particular, our header is available at

```
request.headers["HTTP_GOVUK_ACCOUNT_SESSION"]
```

However, account-api has code to look up the header under a different name, because we weren't previously sure of the correct behaviour.

Now that we know what the correct behaviour is, the code for the incorrect case is redundant and can be removed.

[1]: https://trello.com/c/6GWYTVMM/681-remove-the-redundant-header-logic-in-finder-frontend-and-account-api

[2]: https://github.com/alphagov/frontend/pull/2668