Encrypt & sign session identifiers
We're going to introduce a new level of authentication (LoA) to the
account, and so we want to track which LoA a user logged in with so
that the account-api can reject requests where the user isn't
authenticated enough.

This is so that we can:

- require that the user has done (or had remembered) MFA when they use
  the Transition Checker;

- but NOT require MFA to save pages, where it's too heavyweight.

We want to store the LoA in the session header, so we don't need to
query the auth service on every request.  But we don't want the user
to be able to tamper with it, so we now need to sign the header.

For future-proofing, I've gone for storing a JSON object in the
session, rather than a comma-separated string.  It'll be easier to add
new fields this way.

The encoding is:

    (1) urlsafebase64(
    (2)   encrypt(
            {
              access_token: "access_token",
              refresh_token: "refresh_token",
    (3)     }.to_json,
          )
        )

1. We're storing this in a header, so it needs to be a URL-safe value
2. It's encrypted & signed
3. We turn the hash into a JSON string, because encryption operates on
   strings

To avoid booting out all current users, the legacy unsigned sessions
are accepted for now (and are upgraded into signed sessions).  Those
will be removed once we need to have LoA in the session.

Note: the account-api checking LoA is NOT how LoA will be done for
attributes in the Digital Identity-run attribute store.  If they want
to have attributes which need different LoAs to access, they will
implement that themselves.  It would be a security breach waiting to
happen if they asked every service to independently implement access
checks.  This change is *just* for GOV.UK attributes.

But later cards about handling 403 errors and requesting a specific
LoA when authenticating do move us towards being able to work with a
multi-LoA Digital Identity-run attribute store.