Store the level of authentication in the session
This should be deployed after the change to the account manager, as it
throws an error if the expected level_of_authentication parameter
isn't returned.

The use of LOWEST_LEVEL_OF_AUTHENTICATION in the SessionHeaderHelper
means that current sessions will be upgraded to level0 sessions.  This
won't be enough to use the Transition Checker, so they will end up
needing to reauthenticate, but it is enough to use the bookmark
experiment.