Load seed data before tests
Some tests require a User to be present in the database. We create a
Test User in the seeds file but this is not automatically executed
before running the tests.

This config change populates the test database with seed data before
running tests so that the Test User is present.