Store the user ID in the session
When we start storing information in the account-api database, like
attributes or saved pages, we'll need an identifier to save them
under.  In preparation, this PR stores the user ID from the OIDC ID
Token in the session.

To upgrade existing sessions, the OIDC UserInfo endpoint is called to
fetch the ID.