Return nil when a session has expired tokens and cannot fetch the ID
If a session is sufficiently old, it may have expired tokens in it,
which cause the UserInfo request to fail.

Elsewhere in the app we catch the OAuth failure and return a 401
response (logging the user out) so, to be consistent with that, if the
user's session has expired tokens then log them out.