Create a concern to handle common API functionality
All authenticated APIs need to respond to an OAuth failure with a
401 (because it means the user's session is invalid) and to include
the serialised session identifier in the response body.

Doing that in each controller method results in a lot of duplication,
and also confusion, as logic like this is subtly wrong:

    render json: {
      govuk_account_session: @govuk_account_session.serialise,
      foo: @govuk_account_session.go_something_oauthy,
    }

As the serialisation needs to happen *after* any other calls.