Push local / remote attribute partitioning into the AccountSession
It's error prone to have logic like this:

    foos, bars = partition things
    do_foo_thing foos
    do_bar_thing bars

Especially if there are multiple lines of code between the
partitioning and the usage, where it's easy to lose sight of what
things are supposed to be.  Good naming helps with this.  But why not
eliminate the problem entirely by not requiring whoever is writing the
controller to remember to do this?

Moving the logic into the AccountSession, and replacing the
get_(local|remote)_attributes / set_(local|remote)_attributes with the
higher level get_attributes / set_attributes removes the potential to
mess up in a controller.