Create SavedPage with with OidcUser association (#69)
We need a new table in the account API that is associated with an OIDC
user, so we can store and recall saved page values for a given user.

This feature is very specific to GOV.UK and has little re-use value, so
it is not a candidate for x-gov attributes.

We did consider adding it as a GOV.UK specific attribute. However
attribtues are simple key: value pairs. In this case we want the ability
to return a collection scoped to a user (i.e an index of all a user's
saved pages) and also query / remove individual saved page values.

To do query individual records saved page records as an attribute, we
would have likely had to do something where we returned a blob of all
and then filtered them down.

A table in rails seemed like a better option

Squashed: Create Saved Pages table (#69)

Squashed: That way for a given user, you cannot have duplicate page path values.

Squashed: Create SavePage and associate with OidcUser (#69)

Squashed: Create Factories for OidcUser and SavedPage (#69)