Add routes as resources for saved_pages (#69)
We only need a way to:
- view all (index)
- show a single page (show)
- create a single page (update)*
- delete a single page (destroy)

We don't need an update for the moment as there's barely anything to
these records, and they will only be present initially on a toggle
button, and a place a user can view all their saved pages.

By adding them as a resource we are enforcing that we will be using
restful route design.

If we tried to use different route names or pluralisation rails would be
complain loudly.

Rails resources default to using a path variable called :id to identify
a resouce.

Here however, we are identifying saved pages by their page path, which
is acting as a kind of id. To avoid confusion with the ActiveRecord ID i
have [overriden the named route parameter][1] and called it `page_path`.

You can see this in action by running `bundle exec rails routes` which
will show that this resource needs a page path, not an id.

[1]: https://guides.rubyonrails.org/routing.html#overriding-named-route-parameters

Also we're not using an ID here, but instead uniquely identifying a
resource by a page path. So lets name that here.

* Why are we using Update to create a resource?

We initially had this as a POST endpoint, however after looking around
for similar APIs at GDS we found the [GOV.UK content API][1].

That endpoint uses a PUT method to create or update a resource based on
an ID. This seems closerto what we are doing here. So we decided PUT
makes mores sense than POST.

[1]: https://github.com/alphagov/publishing-api/blob/main/docs/api.md#put-v2contentcontent_id