Add Saved Page endpoint - Create a saved page (#69)
A saved page feature will generally require a route for a user to save a
page. This bit allows for that.

We're using 422 unprocessable_entity to signal if a user tries to save a
page that has already been added. I chose it mainly because it was the
closest HTTP status code I could find to "we've already got one of
those".

A requesting frontend app can choose to handle the response.

We are also specifically handled for a nil or missing page path value as a
bad request.

We only return the value if it persists correctly.

Initial vesions of this also rejected duplicates, however as we are just
posting a page_path which is effectively an ID, we can cheat a bit on
RESTful desing and use find_or_create_by to do an upsert.

So we either return a found record, or we create it.

Earlier drafts had this as POST endpoint, similar to the email
subscriptions API. However as the endpoint is now idempotent, neither
changing the database if a value exists nor rejecting the request
(upsert), then PUT becomes more appropriate.

See discussions here:

- https://stackoverflow.com/questions/630453/what-is-the-difference-between-post-and-put-in-http
- https://stackoverflow.com/questions/18470588/in-rest-is-post-or-put-best-suited-for-upsert-operation