Add Saved Page endpoint - Delete a saved page (#69)
If we are allowing a user to create a saved page record, it seems polite
to let them delete one.

We 404 if we do not have a record that matches the requested value for
the use.

Success is a 204 No Content, once it has been deleted there does not
seem much more to say.

Any other problems with cause `destroy!` to throw an error and we'll get
a 500 devs can investigate