Add Saved Page endpoint - View a saved page (#69)
We are considering adding a button that toggles state based on if the
user has saved the page they are currently on.

I that case we do not want to have to request the index and then parse
through however many values are returned.

This gives an endpoint to request just one page, and find out if it
exists.

If it does we return 200 OK and a JSON body confirming the page path.

If we do not know about it we return a 404