Add validations to SavedPage model (#69)
Prior to this commit we only had database constraints to put limits on
what a user could store here. This left us trying to rescue ActiveRecord
errors as a form of flow control.

A better plan is to user ActiveRecord model validations where we need
them and then use those to return meaningful errors if we could persist
a save page value.

- A page path cannot be nil
- A page path cannot be an empty string

Even the root path of gov.uk will be "/", so empty values are likely
errors and risk duplicates. So we will validate against them.

- A page path cannot be a duplicate scoped to a user

There is little point in a user having the same page saved multiple
times, so let's prevent that from happening. This also sits over a
database index constraint.

- Only a page path

So we don't want:

- protocols (for example: http / https)
- domains (for example: gov.uk) - these will likely get in the way of
testing on our local docker setup too
- query string parameters (for example: /foo?bar) - we use GA tracking
params on parts of the site, might be a bad idea to persist those and
have them remerge unbidden for a later visit
- anchor tags / fragment identifiers (for example: /foo#heading1) we are
saving pages not parts of pages, so we should reject an attempt to jump
to a specific heading that is not guaranteed to be there if there is a
content change.