Add content_id and title to the SavedPage model
We'll need the content ID to handle publishing updates, as not all of
these have a base path (eg, a "vanish" update just specifies the
content ID).  So we can fetch this when the user saves a page (and
throw an error if the content item doesn't exist, rather than let them
save a non-existent page), and use it later in the message queue
consumer.

We're likely to want page titles on the page which lists a user's
saved pages, so I've pulled that in here too.