Fetch the content item when saving a page & store ID + title
If the content item doesn't exist, return a 404.  If it does exist but
has been redirected or marked "gone", then return a 410.  I decided to
go for this over following the redirect automatically, as a redirected
page shouldn't render a save button on the original URL, it should
redirect the original URL.