Implement the message queue processor
For "gone" notifications, redirect if there is an alternative path or
delete the saved page if not.

For "redirect" notifications, redirect if there is a matching
redirect (including gracefully handling prefix redirects) or delete
the saved page if not.

For "vanish" notifications, delete the saved page.

For other notifications, update the saved information.