Cache the OIDC discovery response
We use OIDC discovery to avoid hard-coding things like URLs or keys in
this app.  This method makes an HTTP request to the Identity
Provider (govuk-account-manager-prototype) and then validates the
response.  But this response changes very infrequently, so this is
adding an unnecessary (and slow) network hop to almost every
account-api response.

This is most visible in the auth endpoint which, despite only
generating a URL, commonly takes 150ms because it's got this extra
request and validation going on, but it impacts the attribute and
Transition Checker email endpoints too.