Cache transition_checker_state locally
At some point, we'll move the "Your Account" page from the
account-manager and onto GOV.UK proper (www.gov.uk/account ?).  When
that happens, this account-api will the only user of the
transition_checker_state attribute.

So to improve performance, we'd like to migrate the attribute from the
attribute-service into this app's database.  But we don't want to have
any downtime, so we can't go for the simple approach of:

1. Block updates to the attribute
2. Dump the attribute-service database
3. Import the attribute-service database into the account-api
4. Unblock updates to the attribute

As while the dump&import happens users won't be able to make changes.
Instead we want an online approach, which this commit enables:

1. Get account-api persisting transition_checker_state both
locally *and* remotely.

This is safe because account-api is the only *writer* of this
attribute value.  We don't need to keep anything else in sync.

Then later, we can do this:

2. Add an endpoint to attribute-service to expose the attribute values
3. Run a background task in account-api to fetch values which haven't
already been copied across by (1)
4. Stop account-api persisting the values remotely

This commit by itself is only the first step of the migration.  For
now, it will improve performance for users returning to their saved
results because, as values get cached locally, we won't have to query
the attribute-service to retrieve someone's saved results.