Enforce levels of authentication
We've not yet enabled the ability to log in or register at level0 in
the account manager, so this won't impose any visible changes for
users who have logged in since late April.  Users who logged in before
then and still have a live session will be prompted to reauthenticate
when they next use the Transition Checker.