Move model validations into separate validators
Giving these names makes it clearer what the difference between the
two is.  I haven't added separate specs for these, as they're covered
by the model and request specs.

I changed

    return if redirect_path.nil?
    return if redirect_path.empty?

To

    return if redirect_path.blank?

The only change in behaviour for strings is that a string consisting
only of non-empty whitespace is `blank?` but not `nil?` or `empty?`.
But whtiespace-only values were rejected by this validation
previously, due to not having a leading `/`.