[NEVER MERGE THIS] Test branch for SLO violation on integration
This is a branch that is deliberately broken.

It will be deployed temporarily to integration to drive up our error
count and test the SLO.

If we see alerts as we expect there, then we'll roll back to `main` and
delete this branch forever.

If you find this commit in `main` or on any production server. Something
has gone terrible wrong, rollback over this commit ASAP.