Return the user ID in the `GET /api/user` response
We'll need this in email-alert-api so we can record which GOV.UK
Account a Subscriber belongs to.  I thought about adding this to the
`GET /api/attributes` endpoint, but it feels like more of a user
concern to me.

We might want to retire the attributes endpoints at some point anyway
and consolidate stuff under `/api/user`.