Update linked subscribers on email address change
Currently, if a user updates their email address, any subscriptions
which account-api knows about will also be updated.  But any other
subscriptions for the same email address won't.  This isn't great.

To address this problem, we've added a concept of "linking" a
subscriber to a GOV.UK Account in email-alert-api.

So now the logic for updating email subscriptions is:

- Check if there is a linked subscriber:
  - If so, update the address for that subscriber
  - If not, update each subscription we know about

We're going to make some frontend changes to nudge people towards
linking their accounts.