Use new transition checker endpoint semantics
We changed the account manager to return subscription details (in [1])
as well as a status code.  Since we're going to start using that
response body soon to migrate subscriptions into the account-api
database as they get updated, we need to change our OidcClient and
AccountSession to return the bodies, for use by the controllers.

I've also renamed some methods, to make it clearer that they're
specifically for the Transition Checker subscription.

[1]: https://github.com/alphagov/govuk-account-manager-prototype/pull/885