Handle transition checker subscription in the main email controller
This is so we can update finder-frontend to use what will be the new
email endpoints.  Once we remove the JWT flow, we can change these
endpoints to migrate data from the account-manager, and then after we
do an import of remaining unmigrated data (because some users won't
log in for a while) we can remove this special-case logic.