Incrementally migrate Transition Checker subscriptions to this app
Previously we managed email subscriptions from the Account Manager
prototype running on the PaaS.  This is something we need to get rid
of before we can migrate to the Digital Identity auth solution,
because they won't want to implement the custom auth flow we built to
support that.

We've stopped sending the subscription data directly from the
Transition Checker to the Account Manager.  All communication now goes
through this controller.  So we can now change this controller to stop
sending new data to the Account Manager, and to pull across existing
data as it's accessed.

The `migrate_transition_checker_email_subscription` marks the data as
migrated in the Account Manager, so it won't be updated when (eg) a
user changes their email address: that'll be done through this Account
API, like any other email subscription.

This change is an incremental migration because it won't affect data
which doesn't get accessed (eg, subscriptions for users who don't
return).  For that, we'll need to run a bulk migration of old data.
But stopping creating new data is a necessary first step.