Record timer metrics from OidcClient
Some of the account-api endpoints have a pretty high variance in
response time.  Since the logic in account-api itself is not too
complex, my hunch is that this is the attribute service or account
manager being slow.  So, record some timer metrics to gather some more
evidence.

If it is the OIDC stuff being slow, then maybe we can add more
instances there.  And if it's not, well, this will still be useful
debugging data to have for future problems.