Use strings for attribute names everywhere
A JSON response gets parsed to a hash with string keys, not symbols.
So if we use symbols for attribute names, they don't match the
userinfo hash.

I considered making the userinfo a HashWithIndifferentAccess, but that
would encourage mixing strings and symbols for attribute names
throughout the code, which could be confusing.  So instead, use
strings for attribute names everywhere.

This change will make the optimisation in 26c945a apply to the
/account/home page.