Move attribute definitions to OpenStruct / ActiveRecord
In the UserAttributes class there's a bunch of validation logic to
make sure that the config file is correct, and then in the specs
there's a bunch of test logic to make sure that the validation logic
works, and THEN we get to the real spec: a short one which validates
the config file.

There's already a way to define validations and test them:
ActiveRecord and shoulda-matchers.  And this sort of code is much more
familiar, so it's easier to read at a glance.

So, switch to using an OpenStruct including ActiveRecord::Validations
for attribute definitions, write some (short!) specs around the
validations, just to make sure nobody in the future accidentally
removes one, and then remove all the hand-written validation logic
from UserAttributes.