Set up sidekiq
The app is doing a lot of garbage collection, and specifically it
looks like it's allocating a lot of ActiveRecord transaction objects.

I suspect that that's caused by this in the `sign_in` controller:

    AuthRequest.expired.delete_all

Deleting expired objects on every request was implemented as an easy
"is this really worth setting up sidekiq for?" approach but, if it's
causing excessive allocation, and so excessive garbage collection,
it *is* worth setting up sidekiq to handle.

ActiveJob was already present in the app but I've removed it
completely (rather than configure it to use sidekiq as a queuing
system) because on GOV.UK we're more experienced with using sidekiq
directly.

See also https://github.com/alphagov/govuk-puppet/pull/11200