Merge the `LocalAttribute`s into the `OidcUser` model
Having a separate table with named attributes works well when we don't
know what attributes we're going to have, as in principle it means a
client can call our API with any attribute names and we'd store or
return them appropriately.  But then we undermined that by defining
them all in a config file anyway.

Given that we're not really making use of the flexibility of the
current model, why not go all the way and merge it into a single
database table?  This will reduce database queries and lessen memory
allocation.

This commit will gradually migrate attributes into the `OidcUser`
model as they get updated.  We can do this in a big-bang style by
running this in a rails console:

    OidcUser.find_each(&:set_local_attributes)

Then, when all the `LocalAttribute`s are migrated, we can:

1. Deploy a PR which removes all reference to `LocalAttribute`
2. Then deploy a PR which drops the database table

In the end, `get` and `set_local_attributes` will be:

    def get_local_attributes(names = [])
      names.index_with { |name| self[name] }.compact
    end

    def set_local_attributes(values)
      update!(values)
    end

...or maybe we'll even drop those methods entirely.