Add block to avoid detailed exceptions in Rspec (#154)
As part of the review for this PR [Al noted][1] that the rescue clause
did not need to be there, as `config.action_dispatch.rescue_responses`
allows rails to interpret ActiveRecord errors and provide a sensible
response to avoid a wall of 500 errors.

See [Configuring Rails Apps][2].

However, this gave us Rspec problems in commit 0a6c363

Turns out, request specs are not 'true' full-stack tests. So here Rails
is trying to be helpful and intercepting the error to show you the full
details in test mode, rather than giving a true reflection of what the
user would see in production.

This is causing it to show us the exception, where we just want action
dispatch to do it's thing and 404.

We could change this to be a feature spec instead as one solution.
However Al also [linked to a great thread on rspec-rails][3], that
discusses adding a module containing a block method that wraps a
particular test in specific config to test without detailed exceptions.

This seems like a really nice approach, the method name is clear and
explicit about why we might be seeing some non-standard Request Spec
behaviour here, but allows us to test this path alongside other HTTP
requests.

So that's the solution I've gone with here.

[1]: https://github.com/alphagov/account-api/pull/154#discussion_r674972859
[2]: https://guides.rubyonrails.org/configuring.html#configuring-action-dispatch
[3]: https://github.com/rspec/rspec-rails/issues/2024#issuecomment-420646336