Set up email delivery with Notify
We will use this to send special account-related emails to people;
this functionality won't be directly exposed in an API which other
apps can call to send one-off emails.

We're doing this in account-api rather than email-alert-api because
email-alert-api doesn't do emails to individual addresses: it does
emails to *lists*.  We could emulate this by creating a
single-subscriber list, sending the email, and then deleting the list;
but that's a bit of an abuse, we shouldn't go down that route.

I've gone for an approach similar to email-alert-api[1]: a Sidekiq job
which calls the Notify API directly.

I also looked at the mail-notify gem[2], but it looks like that
doesn't send email asynchronously, so we'd need to wrap calls to it in
a Sidekiq job anyway.

Finally, I considered using ActionMailer directly, like the Account
Manager does[3], but I think that's quite tightly tied to ActiveJob,
and previously I had decided not to introduce ActiveJob to this
project as we don't generally use it in other GOV.UK apps.

So overall, I think a Sidekiq worker which we call with an address,
subject, and body, is the right combination of simplicity and
familiarity.

[1] https://github.com/alphagov/email-alert-api/blob/main/app/workers/send_email_worker.rb
[2] https://github.com/dxw/mail-notify
[3] https://github.com/alphagov/govuk-account-manager-prototype/blob/main/app/mailers/application_mailer.rb