Fetch email & email_verified from OidcUser in EmailSubscription
The `reactivate_if_confirmed!` method needs the `email` and
`email_verified` attributes, which are cached attributes: they may not
be stored locally but, given a user session, we know how to retrieve
them.

So this method used to take the attributes as parameters.

However, this gives a slightly weird API where, strictly speaking,
subscriptions can be associated with a user which use a different
email address, or which are confirmed when the user is not (or vice
versa).  So, to rule out that possibility, treat the attributes as
local attributes in `reactivate_if_confirmed!`, and we just have to
make sure that they are present before calling it.