Implement Transition Checker onboarding email
When a user first signs up to Transition Checker notifications, we
want to send them an onboarding email.

This is part of the unbodging: the Account Manager is now sending a
much more generic email and so, on the GOV.UK side, we want to send
the more specific one.

This email will also be sent if the user signs up and only later uses
the Checker, whereas before they'd just have got the one
Checker-mentioning email regardless of whether they'd used the Checker
or not.

I thought about implementing the logic with some sort of check based
on attribute values, but decided it would be clearer (and less
fragile) to add a `has_received_transition_checker_onboarding_email`
field to the `OidcUser` model.  This new field defaults to `true`, so
we won't start sending any emails yet.  This lets us check the Notify
API keys and suchlike work in every environment, then we can deploy
another migration changing the default to `false` and begin sending
emails.