Make work with both Digital Identity and our Account Manager
This commit changes `OidcClient` and `AccountSession` to work with
Digital Identity and moves Account Manager-specific functionality to a
subclass.  This is because Digital Identity so far has a subset of the
functionality.

There are three APIs Digital Identity are currently lacking:

- **Getting the user's level of authentication, GA session ID, and
cookie consent settings:** we'll need equivalents of these (perhaps
the cookie consent becomes an attribute in the account-api).

- **Getting the value of an attribute:** all the attributes are
returned in the UserInfo response, so we can make do without this.

- **Updating the value of an attribute:** there's no equivalent for
this yet but we don't currently update any remote attributes---we
treat them all as cached---so we can also make do without this.