Refactor: remove `OidcUser.get_local_attributes`
In the places where the attributes are known, use the fields on the
user model instead (eg `u.get_local_attributes(%w[email])` becomes
`u.email`).

In the places where the attributes are not known (they're specified at
runtime), keep using this method, but I've renamed it to make it
clearer that it's not doing anything more special than looking them up
by name.