Add `GET /api/oauth2/end-session` endpoint
This returns the URL to redirect the user to to end their session with
the identity provider (and to trigger single-sign-out if it supports
that).

This will be used in the frontend SessionsController[1], which will
become:

```ruby
def delete
  logout!
  if params[:continue]
    # as it currently is
  elsif params[:done]
    # as it currently is
  else
    redirect_with_ga GdsApi.account_api.get_end_session_url["end_session_uri"]
  end
end
```

When we retire the account manager in production, we can remove the
`params[:continue]` / `params[:done]` cases.

[1] https://github.com/alphagov/frontend/blob/5d7f08f7ad87033f3d602d378609aa6e74d720a0/app/controllers/sessions_controller.rb#L29-L38