Store the ID token JWT in the govuk_account_session
We will use this with the Digital Identity `end_session_endpoint`[1]
for single-sign-out.

I've made this an optional field for now: we have current user
sessions which won't have this, and the `id_token_hint` parameter for
the `end_session_endpoint` is optional.

We may want to make this required in the future, invalidating any old
sessions without it, as it is required if the
`post_logout_redirect_uri` parameter is given.

[1] https://openid.net/specs/openid-connect-session-1_0.html