Add a rake task to publish special routes
So far, these special routes have lived in special-route-publisher.
But we're going to publish the sorting hat page from account-api, as
it won't be a special route (it'll be a help page with some fancy
frontend).

When we have the content and frontend for the sorting hat page ready
to do, we'll add it to this `PublishingApiTasks` class, along with a
rake task.

Since it's potentially confusing having some accounts routes published
from one app and other accounts routes published by this app, I've
decided to move the existing special routes here.  When this commit is
merged, the config for these can be deleted from
https://github.com/alphagov/special-route-publisher/blob/master/data/special_routes.yaml

I've not moved over the `/transition-check` routes, as some of those
are account-related and others are not, and it's probably best to keep
all the transition checker routes together.  The only routes we should
publish from account-api are the ones which are related to core
account functionality.