Make find_or_create OidcUser atomic
The `find_or_create_by!` method isn't atomic, it's implemented like
so:

    def self.find_or_create_by!(options)
      find_by(options) || create!(options)
    end

So if two threads do a `find_or_create_by!` concurrently, there's a
race condition.  For `OidcUser`s, we can't end up with duplicate
records, as we always do `find_or_create_by!(sub: ...)`, and there's a
unique index on `sub`.  Instead we just get a `RecordNotUnique` error,
which isn't great.

We've only seen this issue crop up once, so it's pretty rare.  But
it'll likely become more common as we get more active users.  So,
handle the case where the record isn't unique by doing a `find_by!`.

Unfortunately, there isn't a good way to test that the race condition
is fixed.  But, since the error is ultimately coming from the
database (as it's a constraint violation) not Ruby code, and as it
seems a safe assumption that Postgres is thread-safe, I'm confident it
is.