Special-case `has_unconfirmed_email` attribute when using DI
DI don't have this attribute, so we never cache a value for it, which
means every page which uses it makes a userinfo request.

That's bad because it adds latency (there's an extra network request
going on) and it imposes a session timeout due to the OAuth access
token expiring (if a user tries to use an invalid token we force them
to reauthenticate).

Since DI do not have users with unconfirmed email addresses, as a
hacky work-around we can hard-code this to false.  When we've migrated
to DI in production *and* removed use of this attribute, we can remove
this special case.