Fix flaky test
This test very rarely fails with an off-by-one in the user ID
stubbed (eg, the request is for user 1 but the stub is for user 2).  I
assume there's some race condition between one test being cleaned up
and the next test being run.

Reading the postgres counter directly is a bit of a smell though so,
since in these tests we don't care so much about the exact user ID,
just use a regex URL pattern for the stub.