Replace levels of authentication in permissions with an MFA flag
While we could migrate directly to the Digital Identity
levels-of-authentication equivalent, we don't really need the full
generality of them.  All we need is two levels: with MFA and without.

So rather than support arbitrary levels of authentication (and leak
the knowledge of these levels outside the app, to client apps) we can
simplify.

Before removing the old parameter from the AuthenticationController,
we'll have to update gds-api-adapters.